pull out stress and ipython tests
